Make "Next" always visible in Estuary's OSD
Thanks to Guilouz and the users on the forum for catching this.

"Previous" on the OSD for Estuary is always visible (I assume because it can be used to "restart" a video), but "Next" has a condition to only show when there is a playlist with more than one item. I think this is a simple error, and it was forgotten that next/prev buttons are also used for chapter skipping, rather than an intentional design decision.